+++ src/themes/bootstrap3.js
@@ -389,7 +389,6 @@ class ThemeBootstrap3 extends Theme {
   setTabPaneAttributes (element, active, id) {
     super.setTabPaneAttributes(element, active, id)
     element.classList.add('tab-pane')
-
     element.classList.toggle('in', active)
     element.classList.toggle('active', active)
   }